Raise MixinNotImplementedError
