+++ src/backend/InvenTree/plugin/builtin/integration/inventree_pricing.py
@@ -1,26 +1,25 @@
-"""Pricing mixin class for supporting pricing data."""
+"""Builtin plugin for providing pricing functionality."""
+
+from django.utils.translation import gettext_lazy as _
+
+import structlog
 
 from common.pricing import PriceRangeTuple
-from plugin import PluginMixinEnum
-from plugin.helpers import MixinNotImplementedError
+from plugin import InvenTreePlugin
+from plugin.mixins import PricingMixin, SettingsMixin
+
+logger = structlog.get_logger('inventree')
 
 
-class PricingMixin:
-    """Mixin class which provides support for pricing functionality.
+class InvenTreePricingPlugin(PricingMixin, SettingsMixin, InvenTreePlugin):
+    """Default InvenTree plugin for pricing functionality."""
 
-    - This plugin class provides stubs for pricing-related features.
-    - A default implementation is provided which can be overridden by the plugin.
-    """
-
-    class MixinMeta:
-        """Meta options for this mixin class."""
-
-        MIXIN_NAME = 'Pricing'
-
-    def __init__(self):
-        """Register the mixin."""
-        super().__init__()
-        self.add_mixin(PluginMixinEnum.PRICING, True, __class__)
+    NAME = 'InvenTreePricingPlugin'
+    SLUG = 'inventree-pricing'
+    AUTHOR = _('InvenTree contributors')
+    TITLE = _('InvenTree Pricing Plugin')
+    DESCRIPTION = _('Provides default pricing integration functionality for InvenTree')
+    VERSION = '1.0.0'
 
     def calculate_part_overall_price_range(
         self, part, *args, **kwargs
@@ -37,9 +36,6 @@ class PricingMixin:
         The default implementation calculates the price range based on
         the overall price range generated by the part's pricing data.
         """
-        raise MixinNotImplementedError(
-            'calculate_part_overall_price_range not implemented'
-        )
 
     def calculate_part_bom_price_range(self, part, *args, **kwargs) -> PriceRangeTuple:
         """Calculate the assembly price range for a given part.
@@ -54,7 +50,6 @@ class PricingMixin:
         The "BOM price range" is the cost of manufacturing the assembly
         from its constituent parts, based on their pricing data.
         """
-        raise MixinNotImplementedError('calculate_part_bom_price_range not implemented')
 
     def calculate_part_purchase_price_range(
         self, part, *args, **kwargs
@@ -70,9 +65,6 @@ class PricingMixin:
 
         The "purchase price range" is the cost to purchase the part from suppliers,
         """
-        raise MixinNotImplementedError(
-            'calculate_part_purchase_price_range not implemented'
-        )
 
     def calculate_part_supplier_price_range(
         self, part, *args, **kwargs
@@ -86,9 +78,6 @@ class PricingMixin:
         Returns:
             A tuple representing the min, max supplier price range for the part
         """
-        raise MixinNotImplementedError(
-            'calculate_part_supplier_price_range not implemented'
-        )
 
     def calculate_part_internal_price_range(
         self, part, *args, **kwargs
@@ -102,9 +91,6 @@ class PricingMixin:
         Returns:
             A tuple representing the min, max internal price range for the part
         """
-        raise MixinNotImplementedError(
-            'calculate_part_internal_price_range not implemented'
-        )
 
     def calculate_part_variant_price_range(
         self, part, *args, **kwargs
@@ -120,9 +106,6 @@ class PricingMixin:
 
         The "variant price range" is the price range for all variants of a template part
         """
-        raise MixinNotImplementedError(
-            'calculate_part_variant_price_range not implemented'
-        )
 
     def calculate_part_sale_price_range(self, part, *args, **kwargs) -> PriceRangeTuple:
         """Calculate the sale price range for a given part.
@@ -134,6 +117,3 @@ class PricingMixin:
         Returns:
             A tuple representing the min, max sale price range for the part
         """
-        raise MixinNotImplementedError(
-            'calculate_part_sale_price_range not implemented'
-        )
